SBERDOMA-616 move chack locks to redis quard add time info
