+++ apps/condo/domains/user/utils/redisGuard.js
@@ -2,6 +2,22 @@ const conf = require('@core/config')
 const IORedis = require('ioredis')
 const REDIS_URL = conf['REDIS_URL']
 const { utc } = require('moment')
+const {
+    ApolloError,
+} = require('apollo-errors')
+const {
+    SMS_FOR_IP_DAY_LIMIT_REACHED,
+    SMS_FOR_PHONE_DAY_LIMIT_REACHED,
+    TOO_MANY_REQUESTS,
+} = require('@condo/domains/user/constants/errors')
+
+const {
+    MAX_SMS_FOR_IP_BY_DAY,
+    MAX_SMS_FOR_PHONE_BY_DAY,
+} = require('@condo/domains/user/constants/common')
+
+const phoneWhiteList = Object.keys(conf.SMS_WHITE_LIST ? JSON.parse(conf.SMS_WHITE_LIST) : {})
+const ipWhiteList = conf.IP_WHITE_LIST ? JSON.parse(conf.IP_WHITE_LIST) : []
 
 class RedisGuard {
 
@@ -9,6 +25,32 @@ class RedisGuard {
         this.db = new IORedis(REDIS_URL)
         this.lockPrefix = 'guard_lock:'
         this.counterPrefix = 'guard_counter:'
+    }
+
+    async checkLock (phone, action) {
+        const isLocked = await this.isLocked(phone, action)
+        if (isLocked) {
+            const timeRemain = await this.lockTimeRemain(phone, action)
+            throw new ApolloError(`${TOO_MANY_REQUESTS}]`, {
+                message: `${TOO_MANY_REQUESTS}] resend timeout not expired`,
+                time_thrown: new Date().toISOString(),
+                data: {
+                    timeRemain,
+                },
+            })
+        }
+    }
+
+    async checkDayLimitCounters (phone, rawIp) {
+        const ip = rawIp.split(':').pop()
+        const byPhoneCounter = await this.incrementDayCounter(phone)
+        if (byPhoneCounter > MAX_SMS_FOR_PHONE_BY_DAY && !phoneWhiteList.includes(phone)) {
+            throw new Error(`${SMS_FOR_PHONE_DAY_LIMIT_REACHED}] too many sms requests for this phone number. Try again tomorrow `)
+        }
+        const byIpCounter = await this.incrementDayCounter(ip)
+        if (byIpCounter > MAX_SMS_FOR_IP_BY_DAY && !ipWhiteList.includes(ip)) {
+            throw new Error(`${SMS_FOR_IP_DAY_LIMIT_REACHED}] too many sms requests from this ip address. Try again tomorrow`)
+        }
     }
 
     // Counter will reset at the start of a day ( or after redis restart )
@@ -48,6 +90,15 @@ class RedisGuard {
     }
 
 }
+
+
+
+
+
+
+
+
+
 
 module.exports = {
     RedisGuard,
